requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
reportlab==4.1.0
openpyxl==3.1.2
gunicorn==21.2.0
python-dotenv==1.0.1

# ─────────────────────────────────────────────────────────────────
# Testing & Quality Assurance - أدوات الاختبار وضمان الجودة
# ─────────────────────────────────────────────────────────────────

# Testing Framework - إطار الاختبار الأساسي
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0

# Code Quality - جودة الكود
flake8==6.1.0
black==23.12.1
isort==5.13.2
pylint==3.0.3
mypy==1.7.1

# Performance & Profiling - الأداء والـ Profiling
memory-profiler==0.61.0

# Logging & Monitoring - السجلات والمراقبة
python-json-logger==2.0.7

# Request Mocking - محاكاة الطلبات
responses==0.24.1

# Database Testing - اختبار قاعدة البيانات
factory-boy==3.3.0

# Documentation - التوثيق
pytest-html==4.1.1
sphinx==7.2.6
